+++ src/app/src/pages/redteam/report/components/Overview.tsx
@@ -9,7 +9,7 @@ import {
   severityDisplayNames,
   type Plugin as PluginType,
 } from '@promptfoo/redteam/constants';
-import { getRiskCategorySeverityMap } from '@promptfoo/redteam/shared_frontend';
+import { getRiskCategorySeverityMap } from '@promptfoo/redteam/sharedFrontend';
 import type { RedteamPluginObject } from '@promptfoo/redteam/types';
 import { useReportStore } from './store';
 import './Overview.css';
