+++ src/app/src/pages/redteam/report/components/TestSuites.tsx
@@ -19,7 +19,7 @@ import {
   subCategoryDescriptions,
   type Plugin,
 } from '@promptfoo/redteam/constants';
-import { getRiskCategorySeverityMap } from '@promptfoo/redteam/shared_frontend';
+import { getRiskCategorySeverityMap } from '@promptfoo/redteam/sharedFrontend';
 import type { RedteamPluginObject } from '@promptfoo/redteam/types';
 import './TestSuites.css';
