+++ src/redteam/sharedFrontend.ts
